Move villager module under units and update imports
