+++ tests/test_missing_hotkeys.py
@@ -31,7 +31,7 @@ sys.modules.setdefault("mss", types.SimpleNamespace(mss=lambda: DummyMSS()))
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 import script.common as common
-import script.villager as villager
+import script.units.villager as villager
 
 
 class TestMissingHotkeys(TestCase):
